flink/q1-job/requirements.txt
apache-flink==2.0.0
requests
pillow
numpy
umsgpack
kafka-python

